[analysis] Split CallGraph into dedicated header
