+++ include/triton/Analysis/Utility.h
@@ -287,152 +287,6 @@ bool shouldUseDistSmem(Attribute srcLayout, Attribute dstLayout);
 /// Create a basic DataFlowSolver with constant and dead code analysis included.
 std::unique_ptr<DataFlowSolver> createDataFlowSolver();
 
-namespace triton {
-
-/// This class represents a call graph for a given ModuleOp and holds
-/// data of type T associated with each FunctionOpInterface.
-template <typename T> class CallGraph {
-public:
-  using FuncDataMapT = DenseMap<FunctionOpInterface, T>;
-
-  /// Constructor that builds the call graph for the given moduleOp.
-  explicit CallGraph(ModuleOp moduleOp) : moduleOp(moduleOp) { build(); }
-
-  /// Walks the call graph and applies the provided update functions
-  /// to the edges and nodes.
-  template <WalkOrder UpdateEdgeOrder = WalkOrder::PreOrder,
-            WalkOrder UpdateNodeOrder = WalkOrder::PreOrder,
-            typename UpdateEdgeFn, typename UpdateNodeFn>
-  void walk(UpdateEdgeFn updateEdgeFn, UpdateNodeFn updateNodeFn) {
-    DenseSet<FunctionOpInterface> visited;
-    for (auto root : roots) {
-      doWalk<UpdateEdgeOrder, UpdateNodeOrder>(root, visited, updateEdgeFn,
-                                               updateNodeFn);
-    }
-  }
-
-  /// Retrieves the data associated with a function
-  T *getFuncData(FunctionOpInterface funcOp) {
-    if (funcMap.count(funcOp)) {
-      return &funcMap[funcOp];
-    }
-    return nullptr;
-  }
-
-  /// Getters
-  ModuleOp getModuleOp() const { return moduleOp; }
-  SmallVector<FunctionOpInterface> getRoots() const { return roots; }
-  size_t getNumFunctions() const { return funcMap.size(); }
-
-  /// Returns true if the given function is a root.
-  bool isRoot(FunctionOpInterface funcOp) const {
-    return llvm::is_contained(roots, funcOp);
-  }
-
-  /// Maps the data and the graph nodes associated with a funcOp to a
-  /// targetFuncOp.
-  template <typename FROM, typename TO>
-  void mapFuncOp(FROM funcOp, TO targetFuncOp) {
-    // Iterate over graph and replace
-    for (auto &kv : graph) {
-      for (auto &edge : kv.second) {
-        if (edge.second == funcOp) {
-          edge.second = targetFuncOp;
-        }
-      }
-    }
-    graph[targetFuncOp] = graph[funcOp];
-    // Replace in roots
-    for (auto it = roots.begin(); it != roots.end(); ++it) {
-      if (*it == funcOp) {
-        *it = targetFuncOp;
-        break;
-      }
-    }
-    // Replace in funcMap
-    funcMap[targetFuncOp] = funcMap[funcOp];
-  }
-
-  /// Maps the graph edges associated with a callOp to a targetCallOp.
-  template <typename FROM, typename TO>
-  void mapCallOp(FROM callOp, TO targetCallOp) {
-    // Iterate over graph and replace
-    for (auto &kv : graph) {
-      for (auto &edge : kv.second) {
-        if (edge.first == callOp) {
-          edge.first = targetCallOp;
-        }
-      }
-    }
-  }
-
-private:
-  void build() {
-    SymbolTableCollection symbolTable;
-    DenseSet<FunctionOpInterface> visited;
-    // Build graph
-    moduleOp.walk([&](Operation *op) {
-      auto caller = op->getParentOfType<FunctionOpInterface>();
-      if (auto callOp = dyn_cast<CallOpInterface>(op)) {
-        auto *callee = callOp.resolveCallableInTable(&symbolTable);
-        auto funcOp = dyn_cast_or_null<FunctionOpInterface>(callee);
-        if (funcOp) {
-          graph[caller].emplace_back(
-              std::pair<CallOpInterface, FunctionOpInterface>(callOp, funcOp));
-          visited.insert(funcOp);
-        }
-      }
-    });
-    // Find roots
-    moduleOp.walk([&](FunctionOpInterface funcOp) {
-      if (!visited.count(funcOp)) {
-        roots.push_back(funcOp);
-      }
-    });
-  }
-
-  template <WalkOrder UpdateEdgeOrder = WalkOrder::PreOrder,
-            WalkOrder UpdateNodeOrder = WalkOrder::PreOrder,
-            typename UpdateEdgeFn, typename UpdateNodeFn>
-  void doWalk(FunctionOpInterface funcOp,
-              DenseSet<FunctionOpInterface> &visited, UpdateEdgeFn updateEdgeFn,
-              UpdateNodeFn updateNodeFn) {
-    if (visited.count(funcOp)) {
-      llvm::report_fatal_error("Cycle detected in call graph");
-    }
-    if constexpr (UpdateNodeOrder == WalkOrder::PreOrder) {
-      updateNodeFn(funcOp);
-    }
-    for (auto [callOp, callee] : graph[funcOp]) {
-      if constexpr (UpdateEdgeOrder == WalkOrder::PreOrder) {
-        updateEdgeFn(callOp, callee);
-      }
-      doWalk<UpdateEdgeOrder, UpdateNodeOrder>(callee, visited, updateEdgeFn,
-                                               updateNodeFn);
-      if constexpr (UpdateEdgeOrder == WalkOrder::PostOrder) {
-        updateEdgeFn(callOp, callee);
-      }
-    }
-    if constexpr (UpdateNodeOrder == WalkOrder::PostOrder) {
-      updateNodeFn(funcOp);
-    }
-    visited.erase(funcOp);
-  }
-
-protected:
-  ModuleOp moduleOp;
-  DenseMap<FunctionOpInterface,
-           SmallVector<std::pair<CallOpInterface, FunctionOpInterface>>>
-      graph;
-  FuncDataMapT funcMap;
-  SmallVector<FunctionOpInterface> roots;
-};
-
-} // namespace triton
-
-// Create a basic DataFlowSolver with constant and dead code analysis included.
-std::unique_ptr<DataFlowSolver> createDataFlowSolver();
-
 bool isCvtDimSync(const triton::LinearLayout &srcLayout,
                   const triton::LinearLayout &dstLayout, StringAttr dim);
